Fixed an initial inhibitory weight bug and new boxside switch experiments with independent inputs possible.
For an insufficient amount of excitatory input, the initial
inhibitory weights would be negative to achieve output firing at the
target rate. So I added a rectifier in the function that determines
the initial inhibitory weight.

The boxside switch experiment with independent inputs creates twice
as many centers as usual and assignes one half to the first box side
and the other half to the second box side. When the rat explores the
entire arena, both sets are active. Otherwise only one.